Don't send email if recipient is blank
